// src/app/models/validaterRefRes.ts
export interface ValidateRefResponse {
  message: string;
  valid: boolean;
  stage: string | object;
  test_taken: false;
  name: string
}

// bio: {
//   name: string;
//   mobile_number: string;
//   gender: string;
//   state_of_origin: { id: number; value: string };
//   marital_status: { id: string; value: string };
//   dependents: string;
// };



// tasks for tomorrow
// re enable the input button after test has been taken